Reviewed-by: jgarcke Approved-by: jgarcke
Added optional start and end dates to publisher relationships.
Added a primary flag to descriptors, for determining which of the descriptors to list first when they have the same source and label.  Otherwise, the category determines which sources and labels take precedence over the others.  If no primary flag is present for two identical source/label descriptors, the first one that the application encounters will be displayed first (meaning it is not required to specify a primary).
Also, a few tweaks fixing bugs missed in testing of the last fix- apparently I never did run that full migration I intended to last time around.