Reviewed-by: PeterCroome, andresin Approved-by: andresin
For issues that aren't approved, display a bar below the issue data but above the cover image/credits with a color-coded status message.  Red and orange came from the index status grid, and I picked yellow for skeletons because it seemed to fit, is reasonably inviting to entice people to do something about it (as opposed to "stop- someone else is handling this!") and obviously a white background just wouldn't have the right effect.  

This does not handle situations where the issue and reservations tables disagree on the index status- there's a separate bug for that (#37) and I'll look into it later.  This should get us closer, though.  Fixing bug #37 may take care of the reserved to no one situation.

Also, we'll probably want to come back later and turn the "please index me" bit on the skeleton message into a link to the OI.