Reviewed-by: ralf.haring Approved-by: ralf.haring
This makes the error tracker open in a new window and automatically fill in the URL field.  When you're on the page for a publisher, imprint, publisher's imprint list, series or issue page, the Summary is also filled in.  The django template behavior of evaluating unset variables to an empty string causes the summary to be blank when we don't pass an error_summary to the template, which means it works fine from the front page, search page or various status pages without a special if statement.