Bugs: 65, 74, 83 Reviewed-by: jgarcke Approved-by: jgarcke
This gets rid of the "*GCD" entries to collect weird / disconnected database entries and instead just deletes them.  As that seems to be what is necessary (see bug 65 in particular).  

This doesn't fix *all* page count problems, but it gets a lot of them.  The rest will be fixed either by editors or in a subsequent python script.