Reviewed-by: ralf.haring Approved-by: ralf.haring
This reworks the advanced search algorithm to do everything in terms of joins.  Previously, it would do some things as subqueries.  Well, sort of.  Actually it would suck the subquery results entirely into memory and pass them to an IN clause.  Which was never intended to be permanent but subqueries weren't implemented in Django when I originally wrote this part.  They are now, but MySQL doesn't handle them all that well.  In particular, triple nested subqueries (which could easily happen in the old code) would have atrocious performance.

Now, it just uses joins for everything.  This makes a lot of the code much simpler and/or more uniform, anyway.

As a result of this and a re-database-indexing change I'm about to send out, I've turned on all the advanced search fields again, with a few exceptions:
* No more imprints checkbox, because we now always match both imprints and publishers
* Country is always searched against publishers, as we determined during New Fun discussions that the field on the Series record is superfluous and probably less reliable (it's text instead of a foreign key).
* Searching against indexers remains turned off because we weren't doing it right and it's complicated to do it right.  Maybe we'll fix it in vieux bois.

There are some notable problems that remain:
* If you search for (for instance) publishers based on a story criteria, you'll get each publisher repeated for as many times as it has stories matching the criteria.  Due to the way the DISTINCT clause and pagination work, this is much harder to fix than it looks.  At least I can't think of a way that doesn't do weird things like give you pages with one row each.
* The actual function of the exact match vs contains vs starts with is completely unclear to the user, and quite possibly still implemented wrong.  I'll look at it more tomorrow after this change gets through review.
* Some usability issues that have always been there remain, like it not being at all obvious how to enter dates or how that really works on various different types of searches.
* I rearranged some fields when I commented them out to make them easier to read in the source file, so they may be in sub-optimal arrangements on the page now.  Easily fixed later, I just want to get them functional again right now.