http://dev.comics.org/reviews/r/180/
Display better help for email, names, languages and interests fields.  Reformat the display options inorder to do so correctly (all seems to work best if help text is close under the field, with a space before the next field).  Refactor and add to show_form tag along the way, getting it out of the credits module finally.

Also, handle GET requests on /logout/ and an assortment of related error conditions that can happen logging in and/or out.

Refactor render_error to redirect as an error page should have the URL of an error page, not the page you were trying to reach.  Will fix other error rendering later as I'd like to get all those calling render_error anyway.

Also use a friendlier name for the New Accounts email.