http://dev.comics.org/reviews/r/375/
This adds the ability to also attach a comment to a specific revision, and makes additions and edits of stories as well as edits of issues in the context of story+edit changesets create that attachement.  Other edits still attache the comment only to the changeset, but in those cases it is unambiguous.